backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
pydantic==2.9.2
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
uvicorn==0.32.0
pytest==8.3.3
httpx==0.27.2

# Phase III AI Chatbot Dependencies
cohere==5.20.5
mcp[cli]>=1.2.0
dateparser==1.2.0
pytz==2024.1